requirements.txt
certifi==2019.11.28
Django==2.2
djangorestframework==3.11.0
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
